requirements.txt
streamlit
openai
python-dotenv
gtts
openai-whisper
pandas
numpy